refactor(oauth): make the OAuth cleanup a one-shot command
Replace clean_oauth_tables_daemon with clean_oauth_tables, which runs a
single pass and exits, so it can be scheduled from the shared cleanup
CronJob instead of its own pod. Drops the loop, the entrypoint verb, and
the daemon-only interval and kill-switch settings.
